--- app/templates/gulpfile.js
@@ -1,28 +1,0 @@
-var gulp = require('gulp');
-var jshint = require('gulp-jshint');
-var jshintReporter = require('jshint-stylish');
-var watch = require('gulp-watch');
-
-/*
- * Create variables for our project paths so we can change in one place
- */
-var paths = {
-	'src':['./models/**/*.js','./routes/**/*.js', 'keystone.js', 'package.json']
-};
-
-
-// gulp lint
-gulp.task('lint', function(){
-	gulp.src(paths.src)
-		.pipe(jshint())
-		.pipe(jshint.reporter(jshintReporter));
-
-});
-
-// gulp watcher for lint
-gulp.task('watch:lint', function () {
-	gulp.src(paths.src)
-		.pipe(watch())
-		.pipe(jshint())
-		.pipe(jshint.reporter(jshintReporter));
-});
